+++ indra/sources/utils.py
@@ -2,30 +2,14 @@
 
 """Processor for remote INDRA JSON files."""
 
-from collections import Counter
-
 import requests
-from typing import Collection, List
+from typing import List
 
-from ..statements import Statement, stmts_from_json
+from ..statements import Statement, print_statement_summary, stmts_from_json
 
 __all__ = [
-    'print_statement_summary',
     'RemoteProcessor',
 ]
-
-
-def print_statement_summary(statements: Collection[Statement]):
-    """Print a summary of the statements."""
-    from tabulate import tabulate
-    print(tabulate(
-        Counter(
-            statement.__class__.__name__
-            for statement in statements
-        ).most_common(),
-        headers=["Statement Type", "Count"],
-        tablefmt='github',
-    ))
 
 
 class RemoteProcessor:
